Move Config struct to dbutils. Remove old datastore imports
