+++ cmd/asset-syncer/server/invalidate-cache.go
@@ -19,7 +19,7 @@ package server
 import (
 	"fmt"
 
-	"github.com/kubeapps/common/datastore"
+	"github.com/kubeapps/kubeapps/pkg/dbutils"
 	log "k8s.io/klog/v2"
 )
 
@@ -28,7 +28,7 @@ func InvalidateCache(serveOpts Config, args []string) error {
 		return fmt.Errorf("This command does not take any arguments (got %v)", len(args))
 	}
 
-	dbConfig := datastore.Config{URL: serveOpts.DatabaseURL, Database: serveOpts.DatabaseName, Username: serveOpts.DatabaseUser, Password: serveOpts.DatabasePassword}
+	dbConfig := dbutils.Config{URL: serveOpts.DatabaseURL, Database: serveOpts.DatabaseName, Username: serveOpts.DatabaseUser, Password: serveOpts.DatabasePassword}
 	kubeappsNamespace := serveOpts.KubeappsNamespace
 	manager, err := newManager(dbConfig, kubeappsNamespace)
 	if err != nil {
