+++ cmd/assetsvc/pkg/utils/postgresql_utils.go
@@ -23,7 +23,6 @@ import (
 	"math"
 	"strings"
 
-	"github.com/kubeapps/common/datastore"
 	"github.com/kubeapps/kubeapps/pkg/chart/models"
 	"github.com/kubeapps/kubeapps/pkg/dbutils"
 	_ "github.com/lib/pq"
@@ -39,7 +38,7 @@ type PostgresAssetManager struct {
 	dbutils.PostgresAssetManagerIface
 }
 
-func NewPGManager(config datastore.Config, kubeappsNamespace string) (AssetManager, error) {
+func NewPGManager(config dbutils.Config, kubeappsNamespace string) (AssetManager, error) {
 	m, err := dbutils.NewPGManager(config, kubeappsNamespace)
 	if err != nil {
 		return nil, err
