+++ pkg/dbutils/postgresql_utils.go
@@ -22,7 +22,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/kubeapps/common/datastore"
 	"github.com/kubeapps/kubeapps/pkg/chart/models"
 )
 
@@ -68,7 +67,7 @@ type PostgresAssetManager struct {
 }
 
 // NewPGManager creates an asset manager for PG
-func NewPGManager(config datastore.Config, kubeappsNamespace string) (*PostgresAssetManager, error) {
+func NewPGManager(config dbutils.Config, kubeappsNamespace string) (*PostgresAssetManager, error) {
 	url := strings.Split(config.URL, ":")
 	if len(url) != 2 {
 		return nil, fmt.Errorf("Can't parse database URL: %s", config.URL)
